server/speaker-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
speechbrain>=1.0.0
pyannote.audio>=3.1
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
